Fix: large lists failing to index
A handful of finding aids in the Lilly Repo have been failing to index
because of notes containing lists that exceed the byte limit that
is set for _ssm (string) type fields.

Duke and other institutions have encountered similar issues. Their
solution was to change the _ssm note fields to _tesim (text).

This fix changes controller and config parameters so that
scopecontent notes will be indexed as _tesim.