Change the scopecontent data type
Scopecontent notes are being changed from the 'string' data type to
'text'. This is to attempt to remedy an issue where several finding aids
are failing to index into ArcLight because they have scopecontent notes
that exceed the 'string' byte limit.

These changes are based on the solution that Duke implemented when they
encountered a similar issue with other note types:
https://gitlab.oit.duke.edu/dul-its/dul-arclight/-/commit/98b4dcbe