Fix fonts rendering in serif typeface after viewing printable EAD
Turbolinks was causing the CSS for printable EADs to remain loded
when returning to the previous page, thus rendering some fonts on the
catalog view pages to render in serif fonts.

Disablling turbo-links for the printable and xml EAD views causes
the pages to load with their own CSS, and prevents the print-only
settings from accidentaly being retained by turob-link loaded pages.