Provide better defaults for the date range facet (#262)
Fixes [AR-91](https://bugs.dlib.indiana.edu/browse/AR-91)

ROOT CAUSE:
Traject is mis-interpreting some dates from EADs and there are a
variety of future dates in the index that are causing the range-limit
gem to fail at autodetecting the correct boundaries.

This change adds default boundaries of Year 0 (CE) to two years in
the future.  It also changes the name of the facet.

To fix the underlying problem, we would need to investigate which items
traject is indexin with future dated years (4xxx, 8xxx, 28xx, etc.)
You can see the indexed terms by viewing the Solr schema live:
http://localhost:8983/solr/#/blacklight-core/schema?field=date_range_sim
If you show all terms (i.e. enter 10000 in the box next to Top-Terms,
the bottom of the list will display the anomalous values which can
be clicked on to identify which documents they're assoiciated with.