Refactor year_range (#269)
NGAO currently pulls in the entire Arclight::YearRange class to make
local overrides.  This makes it difficult to tell which parts of the code
are vanilla Arclight and which are the local overrides.

This change refactors that to create the class Ngao::YearRange to
isolate the local changes.

This change also adds assitional logging and expection handling based
on issues identified in various EAD documents.