Index Years as integers instead of strings
This change allows the range slider to do numerical sorting on Year
ranges.  Previously, the data was being saved to a string field, so
we were getting lexical sorting.

This commit adds the necessary field definitions to Solr and
updates the indexer and catalog controller to use the new integer field.